Add seven hour forecast functionality
Modify the sevenhour.js to create seven, hourly forecast cards. Run a split on the hourly-forecast object in the mockdata. pass that to sevenhour.js. In sevenhour.js run a map over the array that creates cards for each object in the array. The card.js file is modified to capture the time, icon, and the temp from the hourly forecast that is based on an or statement. This will allow the app to capture multiple temps from different arrays.